wip: iterate on adding support for non-inner join
